news-api/requirements.txt
fastapi
uvicorn
mysql-connector-python
requests-html
python-dotenv
